refactor return requirement purpose helper
