--- test/support/seeders/data/primary-purposes.js
@@ -1,60 +1,0 @@
-'use strict'
-
-module.exports = [
-  {
-    id: 'b6bb3b77-cfe8-4f22-8dc9-e92713ca3156',
-    legacyId: 'A',
-    description: 'Agriculture',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: '3019b676-e8c6-4bc1-8336-b7862489af70',
-    legacyId: 'E',
-    description: 'Environmental',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: 'd6259e5e-9bb4-4743-b565-e61ec05afc0a',
-    legacyId: 'I',
-    description: 'Industrial, Commercial And Public Services',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: 'c096d17b-7369-4c6b-a230-fb0bc5109b9c',
-    legacyId: 'M',
-    description: 'Amenity',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: '4bbc388a-38cf-425c-bf8f-d609e6a5f57a',
-    legacyId: 'P',
-    description: 'Production Of Energy',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: '7cc6fa87-2e69-4cb3-9abb-a5700ec49b36',
-    legacyId: 'W',
-    description: 'Water Supply',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: '6ad72bf8-ab8e-4fb6-99b0-f54a7acc3a53',
-    legacyId: 'X',
-    description: 'Impounding',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  },
-  {
-    id: 'a38014a7-b8e3-4c3d-87b4-32ca1616d29b',
-    legacyId: 'C',
-    description: 'Crown And Government',
-    createdAt: new Date('2024-07-22'),
-    updatedAt: new Date('2024-07-22')
-  }
-]
